move common build logic to buildUtils.js
